# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "AMONG API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3065",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/user/": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "로그인 상태 유지",
        "description": "로그인 상태 유지 로그인 필요",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "회원 가입",
        "description": "회원 가입",
        "parameters": [
          {
            "name": "wallet_address",
            "in": "body",
            "description": "지갑 주소",
            "schema": {
              "type": "object",
              "properties": {
                "wallet_address": {
                  "type": "string",
                  "example": "input"
                }
              },
              "required": [
                "wallet_address"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "로그인",
        "description": "로그인",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "로그인 예",
            "schema": {
              "type": "object",
              "properties": {
                "wallet_address": {
                  "type": "string",
                  "example": "input"
                },
                "nickname": {
                  "type": "string",
                  "example": "dd"
                }
              },
              "required": [
                "wallet_address",
                "nickname"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/user/id/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "유저 정보",
        "description": "유저 정보",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/user/profile/nickname": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "닉네임 변경",
        "description": "닉네임 변경",
        "parameters": [
          {
            "name": "nickname",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nickname": {
                  "example": "any"
                }
              }
            },
            "description": "원하는 닉네임 입력"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "User"
        ],
        "summary": "닉네임 가져오기 ",
        "description": "닉네임 가져오기 로그인 필요",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/profile/image": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "프로필 사진 등록",
        "description": "프로필 사진 등록 로그인 필요",
        "parameters": [
          {
            "name": "image",
            "in": "body",
            "description": "프로필 사진 예",
            "schema": {
              "type": "object",
              "properties": {
                "image": {
                  "type": "string",
                  "example": "example.png"
                }
              },
              "required": [
                "image"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/image": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "프로필 사진 저장",
        "description": "",
        "parameters": [
          {
            "name": "image",
            "in": "formData",
            "type": "file",
            "description": "프로필 사진 주소"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "User"
        ],
        "summary": "프로필 사진 가져오기",
        "description": "프로필 사진 가져오기",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}